Terminado
Termine las views
