+++ mi_app/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+import 'features/home/presentation/views/failure.dart';
 import 'features/home/presentation/views/initial.dart';
 import 'features/home/presentation/views/loading.dart';
 
@@ -12,23 +13,7 @@ class MyApp extends StatelessWidget {
   
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        body: Container(
-          decoration: BoxDecoration(
-            image: DecorationImage(image: AssetImage('assets/balon2.webp'), fit: BoxFit.cover )
-          ),
-
-          child: Scaffold(
-            
-
-          ),
-          
-
-
-        ),
-      ),
-    );
+    return Failureview();
     
   }
 }
